refactor: move native modules to specific folders
