+++ android/app/src/main/java/com/weatherapp/MainApplication.java
@@ -8,6 +8,8 @@ import com.facebook.react.ReactPackage;
 import com.facebook.react.defaults.DefaultNewArchitectureEntryPoint;
 import com.facebook.react.defaults.DefaultReactNativeHost;
 import com.facebook.soloader.SoLoader;
+import com.weatherapp.native_notification.MyNotificationPackage;
+
 import java.util.List;
 
 public class MainApplication extends Application implements ReactApplication {
